move hook out to ui/hooks/useExtensionUpdates.ts
